Update analyze_fvfm.py
convert datatype to uint8

+++ plantcv/plantcv/photosynthesis/analyze_fvfm.py
@@ -30,9 +30,9 @@ def analyze_fvfm(data, mask, bins=256, label="default"):
     # Auto-increment the device counter
     params.device += 1
     #  Extract frames of interest
-    fdark = data.sel(frame_label='fdark').data
-    fmax = data.sel(frame_label='fmax').data
-    fmin = data.sel(frame_label='fmin').data
+    fdark = data.sel(frame_label='fdark').data.astype(np.uint8)
+    fmax = data.sel(frame_label='fmax').data.astype(np.uint8)
+    fmin = data.sel(frame_label='fmin').data.astype(np.uint8)
     mask = mask.astype(np.uint8)
 
     # QC Fdark Image
